Renamed api query manager.
