+++ src/api/query/IAntSingleResultQueryManager.ts
@@ -1,4 +1,4 @@
 import { IEntity } from '../../model/IEntity';
-import { IAntQueryManager } from './IAntQueryManager';
+import { ApiQueryManager } from './api-query-manager';
 
-export interface IAntSingleResultQueryManager<TEntity extends IEntity> extends IAntQueryManager<TEntity, TEntity> {}
+export interface IAntSingleResultQueryManager<TEntity extends IEntity> extends ApiQueryManager<TEntity, TEntity> {}
