+++ src/api/query/api-query-manager.ts
@@ -1,5 +1,5 @@
 import { IEntity } from '../../model/IEntity';
 import { IQueryManager } from '../../persistence/primary/query/IQueryManager';
 
-export interface IAntQueryManager<TEntity extends IEntity, TResult extends TEntity | TEntity[]>
+export interface ApiQueryManager<TEntity extends IEntity, TResult extends TEntity | TEntity[]>
   extends IQueryManager<TEntity, TResult> {}
