+++ src/test/api/AntModelManagerTest.ts
@@ -1,4 +1,4 @@
-import { AntQueryManager } from '../../api/query/AntQueryManager';
+import { AntQueryManager } from '../../api/query/ant-query-manager';
 import { IEntity } from '../../model/IEntity';
 import { Model } from '../../model/Model';
 import { IModelManager } from '../../persistence/primary/IModelManager';
